# swagger/api_doc.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server for an app that provides time-slots for  interviews between candidates and their interviewers.",
    "version": "1.0.0",
    "title": "Manage Interview API",
    "contact": {
      "email": "nstonny@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:3000",
  "basePath": "/api/v1",
...
  "schemes": [
    "http"
  ],
  "securityDefinitions": {
    "Bearer": {
      "description": "For accessing the API a valid JWT token must be passed in all the queries in\nthe 'Authorization' header.\n\n\nA valid JWT token is generated by the API and retourned as answer of a call\nto the route /login giving a valid user & password.\n\n\nThe following syntax must be used in the 'Authorization' header :\n\n    Bearer: xxxxxx.yyyyyyy.zzzzzz\n",
      "type": "apiKey",
      "name": "x-auth",
      "in": "header"
    }
  },
  "paths": {
...
    "/availabilities": {
      "get": {
        "tags": [
          "availabilities"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Returns all availabilities",
        "operationId": "getAvailabilities",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Availabilities retrieved successfully",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    },
    "/availabilities/{availabilityId}": {
      "get": {
        "tags": [
          "availabilities"
        ],
        "security": [
          {
            "Bearer": []
          }
        ],
        "description": "Returns a single availability",
        "operationId": "getAvailabilityById",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "availabilityId",
            "in": "path",
            "description": "ID of availability to return",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Availability found",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          },
          "401": {
            "description": "Unauthorized",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          },
          "404": {
            "description": "Not found",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/ApiResponse"
            }
          }
        }
      }
    }
  },
  "definitions": {
...
    "ApiResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean"
        },
        "code": {
          "type": "integer"
        },
        "result": {
          "type": "array",
          "description": "List of resulting data",
          "items": {
            "type": "object"
          }
        }
      }
    }
  }
}